+++ src/game/nlcore/elements/control.ts
@@ -2,8 +2,10 @@ import {Actionable} from "@core/action/actionable";
 import {LogicAction} from "@core/action/logicAction";
 import {ContentNode} from "@core/action/tree/actionTree";
 import {Awaitable, Values} from "@lib/util/data";
-import {Chained, ChainedActions, Proxied} from "@core/action/chain";
+import {Chained, Proxied} from "@core/action/chain";
 import {ControlAction} from "@core/action/actions/controlAction";
+import { ActionStatements } from "./type";
+import { Narrator } from "./character";
 
 
 /**@internal */
@@ -19,7 +21,7 @@ export class Control extends Actionable {
      * Execute actions in order, waiting for each action to complete
      * @chainable
      */
-    public static do(actions: ChainedActions): ChainedControl {
+    public static do(actions: ActionStatements): ChainedControl {
         return new Control().do(actions);
     }
 
@@ -27,7 +29,7 @@ export class Control extends Actionable {
      * Execute actions in order, do not wait for this action to complete
      * @chainable
      */
-    public static doAsync(actions: ChainedActions): ChainedControl {
+    public static doAsync(actions: ActionStatements): ChainedControl {
         return new Control().doAsync(actions);
     }
 
@@ -35,7 +37,7 @@ export class Control extends Actionable {
      * Execute all actions at the same time, waiting for any one action to complete
      * @chainable
      */
-    public static any(actions: ChainedActions): ChainedControl {
+    public static any(actions: ActionStatements): ChainedControl {
         return new Control().any(actions);
     }
 
@@ -43,7 +45,7 @@ export class Control extends Actionable {
      * Execute all actions at the same time, waiting for all actions to complete
      * @chainable
      */
-    public static all(actions: ChainedActions): ChainedControl {
+    public static all(actions: ActionStatements): ChainedControl {
         return new Control().all(actions);
     }
 
@@ -51,7 +53,7 @@ export class Control extends Actionable {
      * Execute all actions at the same time, do not wait for all actions to complete
      * @chainable
      */
-    public static allAsync(actions: ChainedActions): ChainedControl {
+    public static allAsync(actions: ActionStatements): ChainedControl {
         return new Control().allAsync(actions);
     }
 
@@ -59,7 +61,7 @@ export class Control extends Actionable {
      * Execute actions multiple times
      * @chainable
      */
-    public static repeat(times: number, actions: ChainedActions): ChainedControl {
+    public static repeat(times: number, actions: ActionStatements): ChainedControl {
         return new Control().repeat(times, actions);
     }
 
@@ -79,7 +81,7 @@ export class Control extends Actionable {
      * Execute actions in order, waiting for each action to complete
      * @chainable
      */
-    public do(actions: ChainedActions): ChainedControl {
+    public do(actions: ActionStatements): ChainedControl {
         return this.push(ControlAction.ActionTypes.do, actions);
     }
 
@@ -87,7 +89,7 @@ export class Control extends Actionable {
      * Execute actions in order, do not wait for this action to complete
      * @chainable
      */
-    public doAsync(actions: ChainedActions): ChainedControl {
+    public doAsync(actions: ActionStatements): ChainedControl {
         return this.push(ControlAction.ActionTypes.doAsync, actions);
     }
 
@@ -95,7 +97,7 @@ export class Control extends Actionable {
      * Execute all actions at the same time, waiting for any one action to complete
      * @chainable
      */
-    public any(actions: ChainedActions): ChainedControl {
+    public any(actions: ActionStatements): ChainedControl {
         return this.push(ControlAction.ActionTypes.any, actions);
     }
 
@@ -103,7 +105,7 @@ export class Control extends Actionable {
      * Execute all actions at the same time, waiting for all actions to complete
      * @chainable
      */
-    public all(actions: ChainedActions): ChainedControl {
+    public all(actions: ActionStatements): ChainedControl {
         return this.push(ControlAction.ActionTypes.all, actions);
     }
 
@@ -111,7 +113,7 @@ export class Control extends Actionable {
      * Execute all actions at the same time, do not wait for all actions to complete
      * @chainable
      */
-    public allAsync(actions: ChainedActions): ChainedControl {
+    public allAsync(actions: ActionStatements): ChainedControl {
         return this.push(ControlAction.ActionTypes.allAsync, actions);
     }
 
@@ -119,7 +121,7 @@ export class Control extends Actionable {
      * Execute actions multiple times
      * @chainable
      */
-    public repeat(times: number, actions: ChainedActions): ChainedControl {
+    public repeat(times: number, actions: ActionStatements): ChainedControl {
         return this.push(ControlAction.ActionTypes.repeat, actions, times);
     }
 
@@ -134,16 +136,26 @@ export class Control extends Actionable {
     /**@internal */
     private push(
         type: Values<typeof ControlAction.ActionTypes>,
-        actions: ChainedActions,
+        actions: ActionStatements,
         ...args: any[]
     ): ChainedControl {
-        const flatted = Chained.toActions(actions);
+        const flatted = this.narrativeToActions(actions);
         const action = new ControlAction(
             this.chain(),
             type,
             new ContentNode().setContent([this.construct(flatted), ...args])
         );
         return this.chain(action);
+    }
+
+    /**@internal */
+    narrativeToActions(statements: ActionStatements): LogicAction.Actions[] {
+        return statements.flatMap(statement => {
+            if (typeof statement === "string") {
+                return Narrator.say(statement).getActions();
+            }
+            return Chained.toActions([statement]);
+        });
     }
 }