+++ src/game/nlcore/elements/sound.ts
@@ -6,8 +6,14 @@ import {SoundActionContentType, SoundActionTypes} from "@core/action/actionTypes
 import {Chained, Proxied} from "@core/action/chain";
 import {SoundAction} from "@core/action/actions/soundAction";
 import {Config, ConfigConstructor} from "@lib/util/config";
+import { StaticScriptWarning } from "../common/Utils";
 
 type ChainedSound = Proxied<Sound, Chained<LogicAction.Actions>>;
+export enum SoundType {
+    Voice = "voice",
+    Bgm = "bgm",
+    Sound = "sound",
+}
 
 /**@internal */
 export type SoundDataRaw = {
@@ -48,6 +54,11 @@ export interface ISoundUserConfig {
      * @default 0
      */
     seek: number;
+    /**
+     * The type of the sound
+     * @default SoundType.Sound
+     */
+    type: SoundType;
 }
 
 type SoundConfig = {
@@ -55,6 +66,7 @@ type SoundConfig = {
     loop: boolean;
     streaming: boolean;
     seek: number;
+    type: SoundType;
 };
 
 type SoundState = {
@@ -75,6 +87,7 @@ export class Sound extends Actionable<SoundDataRaw, Sound> {
         streaming: false,
         rate: 1,
         seek: 0,
+        type: SoundType.Sound,
     });
 
     /**@internal */
@@ -83,6 +96,7 @@ export class Sound extends Actionable<SoundDataRaw, Sound> {
         loop: false,
         streaming: false,
         seek: 0,
+        type: SoundType.Sound,
     });
 
     /**@internal */
@@ -107,6 +121,26 @@ export class Sound extends Actionable<SoundDataRaw, Sound> {
     }
 
     /**@internal */
+    static isSound(v: any): v is Sound {
+        return v instanceof Sound;
+    }
+
+    public static voice(arg0: Partial<ISoundUserConfig> | string) {
+        const config = typeof arg0 === "string" ? {src: arg0} : arg0;
+        return new Sound({...config, type: SoundType.Voice});
+    }
+
+    public static bgm(arg0: Partial<ISoundUserConfig> | string) {
+        const config = typeof arg0 === "string" ? {src: arg0} : arg0;
+        return new Sound({...config, type: SoundType.Bgm});
+    }
+
+    public static sound(arg0: Partial<ISoundUserConfig> | string) {
+        const config = typeof arg0 === "string" ? {src: arg0} : arg0;
+        return new Sound({...config, type: SoundType.Sound});
+    }
+
+    /**@internal */
     public readonly config: Readonly<SoundConfig>;
     /**@internal */
     public state: SoundState;
@@ -115,6 +149,7 @@ export class Sound extends Actionable<SoundDataRaw, Sound> {
 
     constructor(config?: Partial<ISoundUserConfig>);
     constructor(src?: string);
+    constructor(arg0: Partial<ISoundUserConfig> | string)
     constructor(arg0: Partial<ISoundUserConfig> | string = {}) {
         super();
         const rawConfig = typeof arg0 === "string" ? {src: arg0} : arg0;
@@ -127,12 +162,19 @@ export class Sound extends Actionable<SoundDataRaw, Sound> {
     }
 
     /**
-     * Start playing the sound
+     * Start playing the sound and **wait for it to finish**
      *
      * This action will be resolved when the sound reaches the end
      * @chainable
      */
     public play(duration?: number): ChainedSound {
+        if (this.config.type === SoundType.Bgm) {
+            throw new StaticScriptWarning(
+                `Sound (src: ${this.config.src}) is marked as bgm, but it is being played as a normal sound. \n`
+                + "To prevent unintended behavior, the sound marked as bgm cannot be played using `play()`."
+            );
+        }
+
         return this.pushAction<SoundActionContentType["sound:play"]>(SoundAction.ActionTypes.play, [{
             end: this.state.volume,
             duration: duration || 0,
@@ -212,8 +254,9 @@ export class Sound extends Actionable<SoundDataRaw, Sound> {
     }
 
     /**@internal */
-    override reset() {
+    override reset(): this {
         this.state = this.getInitialState(this.userConfig);
+        return this;
     }
 
     /**@internal */