+++ src/game/player/lib/PageRouter/AnimatePresence.tsx
@@ -1,0 +1,6 @@
+import { FirstParam } from "@lib/util/data";
+import { AnimatePresence } from "framer-motion";
+import { JSX } from "react";
+
+export type AnimatePresenceComponent =
+    (arg0: FirstParam<typeof AnimatePresence>) => JSX.Element;